requirements.txt
obs-websocket-py
zulip
zulip_bots
zulip_botserver
